pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vlm-structgen"
version = "0.1.0"
description = "Multimodal structured-generation training and inference stack."
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = "MIT"
dependencies = [
  "cairosvg>=2.8",
  "numpy>=1.26",
  "pillow>=10.0",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "safetensors>=0.5.0",
  "torch>=2.3",
  "torchvision>=0.18",
  "tqdm>=4.66",
  "transformers>=4.52.0",
  "peft>=0.14.0",
  "wandb>=0.19.0",
  "gradio>=5.22.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.9.0",
]
gpu = [
  "flash-attn>=2.7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
